# services/backend/blue-magma-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for blue magma auth and data",
        "title": "Bluemagma API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
...
        "/api/v1/org/{org_id}/report": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve all reports for a specific organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get all reports for an organization",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ReportResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new report for a specific organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Create a new report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Report data",
                        "name": "report",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ReportRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    }
                }
            }
        },
...
        "/api/v1/org/{org_id}/report/{report_id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve a report by its object ID for a specific organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get a report by its object ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Report Object ID",
                        "name": "report_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReportResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a report's details for a specific organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Update a report by its object ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Report Object ID",
                        "name": "report_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Report data",
                        "name": "report",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a report for a specific organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Delete a report by its object ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "org_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Report Object ID",
                        "name": "report_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "fiber.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.QuestionResponse": {
            "type": "object",
            "properties": {
                "answer": {
                    "description": "The answer text",
                    "type": "string"
                },
                "found_properties": {
                    "description": "List of found properties related to the question",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.FoundPropertyResponse"
                    }
                },
                "object_id": {
                    "type": "string"
                },
                "organization_id": {
                    "description": "Foreign key to Organization",
                    "type": "string"
                },
                "question": {
                    "description": "The question text",
                    "type": "string"
                }
            }
        },
...
        "handlers.ReportRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "description": "Status of the report (e.g., \"draft\", \"finalized\")",
                    "type": "string"
                },
                "template_id": {
                    "description": "Foreign key to ReportTemplate",
                    "type": "string"
                }
            }
        },
        "handlers.ReportResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "object_id": {
                    "type": "string"
                },
                "sections": {
                    "description": "List of sections associated with the report",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ReportSectionResponse"
                    }
                },
                "status": {
                    "description": "Status of the report (e.g., \"draft\", \"finalized\")",
                    "type": "string"
                },
                "template_id": {
                    "description": "Foreign key to ReportTemplate",
                    "type": "string"
                }
            }
        },
...
        "handlers.ReportSectionResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "object_id": {
                    "type": "string"
                },
                "report_id": {
                    "description": "Foreign key to Report",
                    "type": "string"
                },
                "rulings": {
                    "description": "List of rulings associated with the report section",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.RulingResponse"
                    }
                }
            }
        },
...
        "handlers.RulingResponse": {
            "type": "object",
            "properties": {
                "decision": {
                    "description": "\"compliant\" | \"non-compliant\" | \"warning\"",
                    "type": "string"
                },
                "level": {
                    "description": "e.g., \"critical\", \"high",
                    "type": "string"
                },
                "object_id": {
                    "type": "string"
                },
                "organization_id": {
                    "description": "Foreign key to Organization",
                    "type": "string"
                },
                "questions": {
                    "description": "List of questions related to the ruling",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.QuestionResponse"
                    }
                },
                "reasoning": {
                    "type": "string"
                },
                "rule_id": {
                    "description": "FK column in Ruling table",
                    "type": "string"
                },
                "status": {
                    "description": "\"pending\" | \"completed\" | \"failed\"",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and your JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}